third_party/WebKit/LayoutTests/fast/dom/shadow/distribution-for-event-path.html

<!doctype html>
<style> body { white-space: pre; } </style>
<a><b></b></a>
<script>

if (window.testRunner)
    testRunner.dumpAsText();

var a = document.querySelector('a');
var b = document.querySelector('b');
var sr = a.createShadowRoot();
sr.innerHTML = '<content></content>';
var content = sr.firstChild;

var log = [];
var phases = {
  2: 'AT_TARGET',
  3: 'BUBBLING_PHASE',
  1: 'CAPTURING_PHASE',
};

[a, b, sr, content].forEach(function(el) {
  [true, false].forEach(function(b) {
    el.addEventListener('click', function(event) {
      log.push([
        event.target.nodeName,
        event.currentTarget.nodeName,
        phases[event.eventPhase]
      ].join(', '));
    }, b);
  });
});

b.click();
document.body.textContent = log.join('\n');

</script>
