third_party/WebKit/LayoutTests/fast/dom/shadow/distribution-for-event-path-expected.txt

B, A, CAPTURING_PHASE
B, #document-fragment, CAPTURING_PHASE
B, CONTENT, CAPTURING_PHASE
B, B, AT_TARGET
B, B, AT_TARGET
B, CONTENT, BUBBLING_PHASE
B, #document-fragment, BUBBLING_PHASE
B, A, BUBBLING_PHASE
